+++ holidays/calendars/__init__.py
@@ -1,0 +1,57 @@
+#  python-holidays
+#  ---------------
+#  A fast, efficient Python library for generating country, province and state
+#  specific sets of holidays on the fly. It aims to make determining whether a
+#  specific date is a holiday as fast and flexible as possible.
+#
+#  Authors: dr-prodigy <dr.prodigy.github@gmail.com> (c) 2017-2023
+#           ryanss <ryanssdev@icloud.com> (c) 2014-2017
+#  Website: https://github.com/dr-prodigy/python-holidays
+#  License: MIT (see LICENSE file)
+
+from datetime import date
+from datetime import timedelta as td
+
+from holidays.calendars.chinese import _ChineseLuniSolar
+from holidays.calendars.custom import _CustomCalendar
+from holidays.calendars.hebrew import _HebrewLuniSolar
+from holidays.calendars.islamic import _IslamicLunar
+from holidays.calendars.thai import _ThaiLuniSolar
+
+GREGORIAN_CALENDAR = "GREGORIAN_CALENDAR"
+JULIAN_CALENDAR = "JULIAN_CALENDAR"
+
+
+def _get_nth_weekday_from(n: int, weekday: int, from_dt: date) -> date:
+    """
+    Return date of a n-th weekday after (n is positive)
+    or before (n is negative) a specific date
+    (e.g. 1st Monday, 2nd Saturday, etc).
+    """
+
+    if n > 0:
+        delta = (n - 1) * 7 + (weekday - from_dt.weekday()) % 7
+    else:
+        delta = (n + 1) * 7 - (from_dt.weekday() - weekday) % 7
+    return from_dt + td(days=delta)
+
+
+def _get_nth_weekday_of_month(
+    n: int, weekday: int, month: int, year: int
+) -> date:
+    """
+    Return date of n-th weekday of month for a specific year
+    (e.g. 1st Monday of Apr, 2nd Friday of June, etc).
+    If n is negative the countdown starts at the end of month
+    (i.e. -1 is last).
+    """
+
+    if n < 0:
+        month += 1
+        if month > 12:
+            month = 1
+            year += 1
+        start_date = date(year, month, 1) + td(days=-1)
+    else:
+        start_date = date(year, month, 1)
+    return _get_nth_weekday_from(n, weekday, start_date)
