+++ holidays/countries/indonesia.py
@@ -9,8 +9,7 @@
 #  Website: https://github.com/dr-prodigy/python-holidays
 #  License: MIT (see LICENSE file)
 
-from datetime import timedelta as td
-
+from holidays.calendars import _CustomCalendar, _IslamicLunar
 from holidays.constants import JAN, FEB, MAR, APR, MAY, JUN, JUL, AUG, SEP
 from holidays.constants import OCT, NOV, DEC
 from holidays.holiday_base import HolidayBase
@@ -34,7 +33,7 @@ class Indonesia(
 
     country = "ID"
     special_holidays = {
-        # Election Day
+        # Election Day.
         2018: ((JUN, 27, "Hari Pemilihan"),),
         2019: ((APR, 17, "Hari Pemilihan"),),
         2020: ((DEC, 9, "Hari Pemilihan"),),
@@ -44,21 +43,20 @@ class Indonesia(
         ChineseCalendarHolidays.__init__(self)
         ChristianHolidays.__init__(self)
         InternationalHolidays.__init__(self)
-        IslamicHolidays.__init__(self)
+        IslamicHolidays.__init__(self, calendar=IndonesiaIslamicCalendar())
         super().__init__(*args, **kwargs)
 
     def _populate(self, year):
         super()._populate(year)
 
-        # New Year's Day
+        # New Year's Day.
         self._add_new_years_day("Tahun Baru Masehi")
 
-        # Chinese New Year
         if year >= 2003:
+            # Chinese New Year.
             self._add_chinese_new_years_day("Tahun Baru Imlek")
 
         if year >= 1983:
-            # Day of Silence
             # https://en.wikipedia.org/wiki/Nyepi
             dates_obs = {
                 2009: (MAR, 26),
@@ -81,172 +79,29 @@ class Indonesia(
                 2026: (MAR, 19),
             }
             if year in dates_obs:
+                # Day of Silence.
                 self._add_holiday("Hari Suci Nyepi", *dates_obs[year])
 
-        # Eid al-Fitr
-        dates_obs = {
-            2001: ((DEC, 16),),
-            2002: ((DEC, 6),),
-            2003: ((NOV, 25),),
-            2004: ((NOV, 14),),
-            2005: ((NOV, 3),),
-            2006: ((OCT, 24),),
-            2007: ((OCT, 13),),
-            2008: ((OCT, 1),),
-            2009: ((SEP, 20),),
-            2010: ((SEP, 10),),
-            2011: ((AUG, 30),),
-            2012: ((AUG, 19),),
-            2013: ((AUG, 8),),
-            2014: ((JUL, 28),),
-            2015: ((JUL, 17),),
-            2016: ((JUL, 6),),
-            2017: ((JUN, 25),),
-            2018: ((JUN, 15),),
-            2019: ((JUN, 5),),
-            2020: ((MAY, 24),),
-            2021: ((MAY, 13),),
-            2022: ((MAY, 2),),
-            2023: ((APR, 22),),
-        }
-        if year in dates_obs:
-            for date_obs in dates_obs[year]:
-                dt = self._add_holiday("Hari Raya Idul Fitri", *date_obs)
-                self._add_holiday(
-                    "Hari kedua dari Hari Raya Idul Fitri", dt + td(days=+1)
-                )
-        else:
-            self._add_eid_al_fitr_day("Hari Raya Idul Fitri* (*estimated)")
-            self._add_eid_al_fitr_day_two(
-                "Hari kedua dari Hari Raya Idul Fitri* (*estimated)"
-            )
+        # Eid al-Fitr.
+        self._add_eid_al_fitr_day("Hari Raya Idul Fitri")
+        self._add_eid_al_fitr_day_two("Hari kedua dari Hari Raya Idul Fitri")
 
-        # Eid al-Adha
-        dates_obs = {
-            2001: ((MAR, 6),),
-            2002: ((FEB, 23),),
-            2003: ((FEB, 12),),
-            2004: ((FEB, 2),),
-            2005: ((JAN, 21),),
-            2006: ((JAN, 10), (DEC, 31)),
-            2007: ((DEC, 20),),
-            2008: ((DEC, 8),),
-            2009: ((NOV, 27),),
-            2010: ((NOV, 17),),
-            2011: ((NOV, 6),),
-            2012: ((OCT, 26),),
-            2013: ((OCT, 15),),
-            2014: ((OCT, 5),),
-            2015: ((SEP, 24),),
-            2016: ((SEP, 12),),
-            2017: ((SEP, 1),),
-            2018: ((AUG, 22),),
-            2019: ((AUG, 11),),
-            2020: ((JUL, 31),),
-            2021: ((JUL, 20),),
-            2022: ((JUL, 10),),
-            2023: ((JUN, 29),),
-        }
-        if year in dates_obs:
-            for date_obs in dates_obs[year]:
-                self._add_holiday("Hari Raya Idul Adha", *date_obs)
-        else:
-            self._add_eid_al_adha_day("Hari Raya Idul Adha* (*estimated)")
+        # Eid al-Adha.
+        self._add_eid_al_adha_day("Hari Raya Idul Adha")
 
-        # Islamic New Year
-        dates_obs = {
-            2001: ((MAR, 26),),
-            2002: ((MAR, 15),),
-            2003: ((MAR, 5),),
-            2004: ((FEB, 22),),
-            2005: ((FEB, 10),),
-            2006: ((JAN, 31),),
-            2007: ((JAN, 20),),
-            2008: ((JAN, 10), (DEC, 29)),
-            2009: ((DEC, 18),),
-            2010: ((DEC, 7),),
-            2011: ((NOV, 27),),
-            2012: ((NOV, 15),),
-            2013: ((NOV, 5),),
-            2014: ((OCT, 25),),
-            2015: ((OCT, 14),),
-            2016: ((OCT, 2),),
-            2017: ((SEP, 21),),
-            2018: ((SEP, 11),),
-            2019: ((SEP, 1),),
-            2020: ((AUG, 20),),
-            2021: ((AUG, 11),),
-            2022: ((JUL, 30),),
-        }
-        if year in dates_obs:
-            for date_obs in dates_obs[year]:
-                self._add_holiday("Tahun Baru Islam", *date_obs)
-        else:
-            self._add_islamic_new_year_day("Tahun Baru Islam* (*estimated)")
+        # Islamic New Year.
+        self._add_islamic_new_year_day("Tahun Baru Islam")
 
-        # The Prophet's Birthday
-        dates_obs = {
-            2006: ((APR, 10),),
-            2007: ((MAR, 31),),
-            2008: ((MAR, 20),),
-            2009: ((MAR, 9),),
-            2010: ((FEB, 26),),
-            2011: ((FEB, 15),),
-            2012: ((FEB, 5),),
-            2013: ((JAN, 24),),
-            2014: ((JAN, 14),),
-            2015: ((JAN, 3), (DEC, 24)),
-            2016: ((DEC, 12),),
-            2017: ((DEC, 1),),
-            2018: ((NOV, 20),),
-            2019: ((NOV, 9),),
-            2020: ((OCT, 29),),
-            2021: ((OCT, 19),),
-            2022: ((OCT, 8),),
-        }
-        if year in dates_obs:
-            for date_obs in dates_obs[year]:
-                self._add_holiday("Maulid Nabi Muhammad", *date_obs)
-        else:
-            self._add_mawlid_day("Maulid Nabi Muhammad* (*estimated)")
+        # The Prophet's Birthday.
+        self._add_mawlid_day("Maulid Nabi Muhammad")
 
-        # The Prophet's Ascension
-        dates_obs = {
-            2001: ((OCT, 15),),
-            2002: ((OCT, 4),),
-            2003: ((SEP, 24),),
-            2004: ((SEP, 12),),
-            2005: ((SEP, 1),),
-            2006: ((AUG, 22),),
-            2007: ((AUG, 11),),
-            2008: ((JUL, 31),),
-            2009: ((JUL, 20),),
-            2010: ((JUL, 9),),
-            2011: ((JUN, 29),),
-            2012: ((JUN, 17),),
-            2013: ((JUN, 6),),
-            2014: ((MAY, 27),),
-            2015: ((MAY, 16),),
-            2016: ((MAY, 6),),
-            2017: ((APR, 24),),
-            2018: ((APR, 14),),
-            2019: ((APR, 3),),
-            2020: ((MAR, 22),),
-            2021: ((MAR, 11),),
-            2022: ((FEB, 28),),
-        }
-        if year in dates_obs:
-            for date_obs in dates_obs[year]:
-                self._add_holiday("Isra' Mi'raj Nabi Muhammad", *date_obs)
-        else:
-            self._add_isra_and_miraj_day(
-                "Isra' Mi'raj Nabi Muhammad* (*estimated)"
-            )
+        # The Prophet's Ascension.
+        self._add_isra_and_miraj_day("Isra' Mi'raj Nabi Muhammad")
 
-        # Good Friday
+        # Good Friday.
         self._add_good_friday("Wafat Yesus Kristus")
 
-        # Buddha's Birthday
+        # Buddha's Birthday.
         if year >= 1983:
             date_obs = {
                 2007: (JUN, 1),
@@ -268,27 +123,29 @@ class Indonesia(
                 2023: (JUN, 4),
             }
             if year in date_obs:
+                # Buddha's Birthday.
                 self._add_holiday("Hari Raya Waisak", *date_obs[year])
             else:
+                # Buddha's Birthday.
                 self._add_chinese_calendar_holiday(
                     "Hari Raya Waisak* (*estimated)", 4, 15
                 )
 
-        # Labour Day
         if 1953 <= year <= 1968 or year >= 2014:
+            # Labor Day.
             self._add_labor_day("Hari Buruh Internasional")
 
-        # Ascension Day
+        # Ascension Day.
         self._add_ascension_thursday("Kenaikan Yesus Kristus")
 
-        # Pancasila Day
         if year >= 2017:
+            # Pancasila Day.
             self._add_holiday("Hari Lahir Pancasila", JUN, 1)
 
-        # Independence Day
+        # Independence Day.
         self._add_holiday("Hari Kemerdekaan Republik Indonesia", AUG, 17)
 
-        # Christmas Day
+        # Christmas Day.
         self._add_christmas_day("Hari Raya Natal")
 
 
@@ -298,3 +155,127 @@ class ID(Indonesia):
 
 class IDN(Indonesia):
     pass
+
+
+class IndonesiaIslamicCalendar(_CustomCalendar, _IslamicLunar):
+    EID_AL_ADHA_DATES = {
+        2001: ((MAR, 6),),
+        2002: ((FEB, 23),),
+        2003: ((FEB, 12),),
+        2004: ((FEB, 2),),
+        2005: ((JAN, 21),),
+        2006: ((JAN, 10), (DEC, 31)),
+        2007: ((DEC, 20),),
+        2008: ((DEC, 8),),
+        2009: ((NOV, 27),),
+        2010: ((NOV, 17),),
+        2011: ((NOV, 6),),
+        2012: ((OCT, 26),),
+        2013: ((OCT, 15),),
+        2014: ((OCT, 5),),
+        2015: ((SEP, 24),),
+        2016: ((SEP, 12),),
+        2017: ((SEP, 1),),
+        2018: ((AUG, 22),),
+        2019: ((AUG, 11),),
+        2020: ((JUL, 31),),
+        2021: ((JUL, 20),),
+        2022: ((JUL, 10),),
+        2023: ((JUN, 29),),
+    }
+
+    EID_AL_FITR_DATES = {
+        2001: ((DEC, 16),),
+        2002: ((DEC, 6),),
+        2003: ((NOV, 25),),
+        2004: ((NOV, 14),),
+        2005: ((NOV, 3),),
+        2006: ((OCT, 24),),
+        2007: ((OCT, 13),),
+        2008: ((OCT, 1),),
+        2009: ((SEP, 20),),
+        2010: ((SEP, 10),),
+        2011: ((AUG, 30),),
+        2012: ((AUG, 19),),
+        2013: ((AUG, 8),),
+        2014: ((JUL, 28),),
+        2015: ((JUL, 17),),
+        2016: ((JUL, 6),),
+        2017: ((JUN, 25),),
+        2018: ((JUN, 15),),
+        2019: ((JUN, 5),),
+        2020: ((MAY, 24),),
+        2021: ((MAY, 13),),
+        2022: ((MAY, 2),),
+        2023: ((APR, 22),),
+    }
+
+    HIJRI_NEW_YEAR_DATES = {
+        2001: ((MAR, 26),),
+        2002: ((MAR, 15),),
+        2003: ((MAR, 5),),
+        2004: ((FEB, 22),),
+        2005: ((FEB, 10),),
+        2006: ((JAN, 31),),
+        2007: ((JAN, 20),),
+        2008: ((JAN, 10), (DEC, 29)),
+        2009: ((DEC, 18),),
+        2010: ((DEC, 7),),
+        2011: ((NOV, 27),),
+        2012: ((NOV, 15),),
+        2013: ((NOV, 5),),
+        2014: ((OCT, 25),),
+        2015: ((OCT, 14),),
+        2016: ((OCT, 2),),
+        2017: ((SEP, 21),),
+        2018: ((SEP, 11),),
+        2019: ((SEP, 1),),
+        2020: ((AUG, 20),),
+        2021: ((AUG, 11),),
+        2022: ((JUL, 30),),
+    }
+
+    ISRA_AND_MIRAJ_DATES = {
+        2001: ((OCT, 15),),
+        2002: ((OCT, 4),),
+        2003: ((SEP, 24),),
+        2004: ((SEP, 12),),
+        2005: ((SEP, 1),),
+        2006: ((AUG, 22),),
+        2007: ((AUG, 11),),
+        2008: ((JUL, 31),),
+        2009: ((JUL, 20),),
+        2010: ((JUL, 9),),
+        2011: ((JUN, 29),),
+        2012: ((JUN, 17),),
+        2013: ((JUN, 6),),
+        2014: ((MAY, 27),),
+        2015: ((MAY, 16),),
+        2016: ((MAY, 6),),
+        2017: ((APR, 24),),
+        2018: ((APR, 14),),
+        2019: ((APR, 3),),
+        2020: ((MAR, 22),),
+        2021: ((MAR, 11),),
+        2022: ((FEB, 28),),
+    }
+
+    MAWLID_DATES = {
+        2006: ((APR, 10),),
+        2007: ((MAR, 31),),
+        2008: ((MAR, 20),),
+        2009: ((MAR, 9),),
+        2010: ((FEB, 26),),
+        2011: ((FEB, 15),),
+        2012: ((FEB, 5),),
+        2013: ((JAN, 24),),
+        2014: ((JAN, 14),),
+        2015: ((JAN, 3), (DEC, 24)),
+        2016: ((DEC, 12),),
+        2017: ((DEC, 1),),
+        2018: ((NOV, 20),),
+        2019: ((NOV, 9),),
+        2020: ((OCT, 29),),
+        2021: ((OCT, 19),),
+        2022: ((OCT, 8),),
+    }
